+++ pay/src/main/java/com/yp/pay/entity/aliandwx/req/AliWebPayReq.java
@@ -35,14 +35,15 @@ public class AliWebPayReq {
     private String subject;
 
     /**
-     * 订单总金额，单位为元，精确到小数点后两位，取值范围[0.01,100000000] 如果同时传入
+     * 订单总金额，单位为分
+     * 如果同时传入
      * 【可打折金额】和【不可打折金额】，该参数可以不用传入； 如果同时传入了【可打折金额】，
      * 【不可打折金额】，【订单总金额】三者，则必须满足如下条件：
      * 【订单总金额】=【可打折金额】+【不可打折金额】
      */
     @NotNull(message = "支付总金额不能为空")
     @ApiModelProperty(value = "订单总金额", required = true)
-    private BigDecimal totalAmount;
+    private Integer totalAmount;
 
     /**
      * 收款支付宝用户ID。 如果该值为空，则默认为商户签约账号对应的支付宝用户ID
